feat: Add calendar management components and hooks
- Implemented CalendarActionsSection for managing calendar actions (clear, enable, save).
- Created CurrentStateDisplay to show current calendar visibility state.
- Developed ImportExportSection for handling import/export functionality with keyboard shortcuts.
- Added LensIcon component for rendering the extension icon with fallback.
- Introduced PresetSelectionSection for managing calendar presets with update/delete functionality.
- Created hooks for calendar operations, import/export operations, and preset operations with error handling.
- Added utility functions for dropdown styles, notifications, and operation helpers.
