+++ src/presentation/utils/dropdownStyles.ts
@@ -1,0 +1,27 @@
+/**
+ * Dropdown Styles Utility
+ * Reusable styles for dropdown components
+ */
+
+export const dropdownContainerStyle = {
+  background: '#fff',
+  borderRadius: '8px',
+  boxShadow: '0 6px 16px 0 rgba(0, 0, 0, 0.08), 0 3px 6px -4px rgba(0, 0, 0, 0.12), 0 9px 28px 8px rgba(0, 0, 0, 0.05)',
+  padding: '12px',
+  minWidth: '280px'
+};
+
+export const sectionHeaderStyle = {
+  fontSize: '12px',
+  color: '#666',
+  marginBottom: '6px',
+  fontWeight: 500
+};
+
+export const sectionContainerStyle = {
+  marginBottom: '12px'
+};
+
+export const dividerStyle = {
+  margin: '8px 0'
+};
