+++ src/presentation/utils/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Utilities Index
+ * Centralized exports for presentation utilities
+ */
+
+export * from './customNotification';
+export * from './dropdownStyles';
+export * from './notificationHelpers';
+export * from './operationHelpers';
